image preview in solution modal popup

+++ Aephy.WEB.Admin/Controllers/SolutionsController.cs
@@ -45,32 +45,31 @@ namespace Aephy.WEB.Admin.Controllers
             IFormFile imageFile = httpPostedFileBase;
 
             var solutionData = await _apiRepository.MakeApiCallAsync("api/Admin/AddorEditSolutionData", HttpMethod.Post, result);
-            if (result.Id == 0)
+            dynamic data = JsonConvert.DeserializeObject(solutionData);
+            if (data["StatusCode"] == 200)
             {
-                dynamic data = JsonConvert.DeserializeObject(solutionData);
-                if (data != null)
+                if (result.Id == 0)
                 {
+
                     int Id = data.Result;
-                    var d = await SaveImageFile(imageFile, Id);
-                    var ok = await _apiRepository.MakeApiCallAsync("api/Admin/UpdateImage", HttpMethod.Post, d);
+                    var fileData = await SaveImageFile(imageFile, Id);
+                    await _apiRepository.MakeApiCallAsync("api/Admin/UpdateImage", HttpMethod.Post, fileData);
+
                 }
-            }
-            else
-            {
-                dynamic data = JsonConvert.DeserializeObject(solutionData);
-                if (data != null)
+                else
                 {
-                    int Id = result.Id;
-                    string Imagepath = data.Result;
-                    var d = await EditImageFile(imageFile, Id, Imagepath);
-                    var ok = await _apiRepository.MakeApiCallAsync("api/Admin/UpdateImageById", HttpMethod.Post, d);
-                }
+                    if(imageFile != null)
+                    {
+                        int Id = result.Id;
+                        string Imagepath = data.Result;
+                        var editFileData = await EditImageFile(imageFile, Id, Imagepath);
+                        await _apiRepository.MakeApiCallAsync("api/Admin/UpdateImageById", HttpMethod.Post, editFileData);
+                    }
 
+                }
             }
 
-
-
-            return "";
+            return solutionData;
         }
 
 
@@ -157,22 +156,20 @@ namespace Aephy.WEB.Admin.Controllers
             dynamic data = JsonConvert.DeserializeObject(serviceList);
             try
             {
-                if (data.Result != null)
+                if (data["StatusCode"] == 200)
                 {
                     string imagepath = data.Result.ImagePath;
                     string sasToken = GenerateSasToken(imagepath);
                     imageUrlWithSas = $"{data.Result.ImagePath}?{sasToken}";
                     data.Result.ImageUrlWithSas = imageUrlWithSas;
-
                 }
             }
             catch (Exception ex)
             {
-
+                return ex.Message + ex.InnerException;
             }
-            string jsonString = JsonConvert.SerializeObject(data, Formatting.Indented);
-
-            return jsonString;
+            string convertjsonTostring = JsonConvert.SerializeObject(data, Formatting.Indented);
+            return convertjsonTostring;
         }
 
         private static string GetEndpointSuffixFromConnectionString(string connectionString)
@@ -321,35 +318,38 @@ namespace Aephy.WEB.Admin.Controllers
 
         }
 
-        public async Task<EditSolutionImage> EditImageFile(IFormFile imageFile, object Id, string Imagepath)
+        public async Task<SolutionImage> EditImageFile(IFormFile imageFile, object Id, string Imagepath)
         {
-
-            EditSolutionImage solutions = new EditSolutionImage();
-            if (!string.IsNullOrEmpty(Imagepath))
+            SolutionImage solutions = new SolutionImage();
+            if (imageFile != null && imageFile.Length > 0)
             {
-                BlobServiceClient updateBlobServiceClient = new BlobServiceClient(_connectionString);
-                BlobContainerClient updateContainerClient = updateBlobServiceClient.GetBlobContainerClient("profileimages");
+                if (!string.IsNullOrEmpty(Imagepath))
+                {
+                    BlobServiceClient updateBlobServiceClient = new BlobServiceClient(_connectionString);
+                    BlobContainerClient updateContainerClient = updateBlobServiceClient.GetBlobContainerClient("profileimages");
 
-                BlobClient updateBlobClient = updateContainerClient.GetBlobClient(Path.GetFileName(Imagepath));
+                    BlobClient updateBlobClient = updateContainerClient.GetBlobClient(Path.GetFileName(Imagepath));
 
-                await updateBlobClient.DeleteIfExistsAsync();
+                    await updateBlobClient.DeleteIfExistsAsync();
+                }
+
+                string fileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
+
+                BlobServiceClient createBlobServiceClient = new BlobServiceClient(_connectionString);
+                BlobContainerClient createContainerClient = createBlobServiceClient.GetBlobContainerClient("profileimages");
+
+                BlobClient createBlobClient = createContainerClient.GetBlobClient(fileName);
+
+                using (var stream = imageFile.OpenReadStream())
+                {
+                    await createBlobClient.UploadAsync(stream, overwrite: true);
+                }
+
+                Imagepath = createBlobClient.Uri.ToString();
+                solutions.Id = (int)Id;
+                solutions.ImagePath = Imagepath;
             }
-
-            string fileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
-
-            BlobServiceClient createBlobServiceClient = new BlobServiceClient(_connectionString);
-            BlobContainerClient createContainerClient = createBlobServiceClient.GetBlobContainerClient("profileimages");
-
-            BlobClient createBlobClient = createContainerClient.GetBlobClient(fileName);
-
-            using (var stream = imageFile.OpenReadStream())
-            {
-                await createBlobClient.UploadAsync(stream, overwrite: true);
-            }
-
-            Imagepath = createBlobClient.Uri.ToString();
-            solutions.Id = (int)Id;
-            solutions.ImagePath = Imagepath;
+               
             return solutions;
